Clean up some 'code smells' found by SonarQube.  Fix some bugs.   However, the HMAC tests are all failing now, so that still needs to be fixed.

+++ keystorage/database/secretdatabase.h
@@ -39,7 +39,7 @@ private:
     bool queryEntryToString(const QSqlQuery &query, const QString &column, QString &result);
     bool queryEntryToInt(const QSqlQuery &query, const QString &column, int &result);
 
-    bool upgradeSchema(int from);
+    bool upgradeSchema(size_t from);
 
     void copy(const SecretDatabase &toCopy);
